destroyアクションに条件分岐を追加

+++ app/controllers/items_controller.rb
@@ -38,6 +38,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    unless item.user_id == current_user.id
+      redirect_to items_path
+    end
     item.destroy
     redirect_to items_path
   end
